fw/astep24-3l/verification/requirements.txt
cocotbext-uart
pytest
bitstring~=3.1
PyYAML